+++ spec/support/data_helpers.rb
@@ -39,6 +39,1170 @@ module EnergySparksAnalyticsDataHelpers
     AggregateDataService.new(meter_collection).aggregate_heat_and_electricity_meters
     meter_collection
   end
+
+  def benchmark_database_hash
+    {
+      Date.parse('Mon, 06 Feb 2023') => {
+        795 => {
+          'gsoo_sdop' => 0.4031742638136088,
+          'gsoo_sdcp' => 0.37950806268295123,
+          'gsoo_holp' => 0.08583469084374912,
+          'gsoo_wkep' => 0.13148298265969088,
+          'gsoo_comp' => 0.0,
+          'gsoo_ahl£' => 296.61605099999997,
+          'gsoo_awk£' => 454.3613159999998,
+          'gsoo_com£' => 0.0,
+          'gsoo_aoo£' => 2062.430600758286,
+          'gsoo_ahl€' => 296.61605099999997,
+          'gsoo_awk€' => 454.3613159999998,
+          'gsoo_com€' => 0.0,
+          'gsoo_aoo€' => 2062.430600758286,
+          'gsoo_esv€' => 1025.7307020892813,
+          'gsoo_ratg' => 2.6,
+          'hwia_ratg' => 8,
+          'httc_r2' => 0.36845802094728347,
+          'httc_sav€' => 736.146730355791,
+          'httc_ratg' => 4,
+          'gswe_ratg' => 10.0,
+          'enba_klyr' => 115_188.87762988935,
+          'enba_£lyr' => 3455.666328896681,
+          'enba_€lyr' => 3455.666328896681,
+          'enba_co2y' => 24_189.664302276764,
+          'enba_co2t' => 24.189664302276764,
+          'enba_co2e' => nil,
+          'enba_co2g' => 24_189.664302276764,
+          'enba_co2h' => nil,
+          'enba_co2s' => nil,
+          'enba_kpup' => 567.4328947285189,
+          'enba_£pup' => 17.022986841855573,
+          'enba_cpup' => 119.16090789298899,
+          'enba_kfla' => 74.45952012274684,
+          'enba_£fla' => 2.2337856036824055,
+          'enba_cfla' => 15.636499225776834,
+          'enba_pp%d' => 0.2295147099230237,
+          'enba_sact' => Date.parse('Fri, 02 Dec 2022'),
+          'enba_solr' => '',
+          'enba_ke0' => nil,
+          'enba_ken' => nil,
+          'enba_kenp' => nil,
+          'enba_kg0' => 115_188.87762988935,
+          'enba_kgn' => 209_114.63420000012,
+          'enba_kgnp' => -0.4491591749637133,
+          'enba_kh0' => nil,
+          'enba_khn' => nil,
+          'enba_khnp' => nil,
+          'enba_ks0' => nil,
+          'enba_ksn' => nil,
+          'enba_ksnp' => nil,
+          'enba_kx0' => 115_188.87762988935,
+          'enba_kxn' => 209_114.63420000012,
+          'enba_kxnp' => -0.4491591749637133,
+          'enba_ce0' => nil,
+          'enba_cen' => nil,
+          'enba_cenp' => nil,
+          'enba_cg0' => 24_189.664302276764,
+          'enba_cgn' => 43_914.07318199999,
+          'enba_cgnp' => -0.44915917496371294,
+          'enba_ch0' => nil,
+          'enba_chn' => nil,
+          'enba_chnp' => nil,
+          'enba_cs0' => nil,
+          'enba_csn' => nil,
+          'enba_csnp' => nil,
+          'enba_cx0' => 24_189.664302276764,
+          'enba_cxn' => 43_914.07318199999,
+          'enba_cxnp' => -0.44915917496371294,
+          'enba_pe0' => nil,
+          'enba_pen' => nil,
+          'enba_penp' => nil,
+          'enba_pg0' => 3455.666328896681,
+          'enba_pgn' => 6273.439025999994,
+          'enba_pgnp' => -0.4491591749637124,
+          'enba_ph0' => nil,
+          'enba_phn' => nil,
+          'enba_phnp' => nil,
+          'enba_px0' => 3455.666328896681,
+          'enba_pxn' => 6273.439025999994,
+          'enba_pxnp' => -0.4491591749637124,
+          'enba_€e0' => nil,
+          'enba_€en' => nil,
+          'enba_€enp' => nil,
+          'enba_€g0' => 3455.666328896681,
+          'enba_€gn' => 6273.439025999994,
+          'enba_€gnp' => -0.4491591749637124,
+          'enba_€h0' => nil,
+          'enba_€hn' => nil,
+          'enba_€hnp' => nil,
+          'enba_€s0' => nil,
+          'enba_€sn' => nil,
+          'enba_€snp' => nil,
+          'enba_€x0' => 3455.666328896681,
+          'enba_€xn' => 6273.439025999994,
+          'enba_€xnp' => -0.4491591749637124,
+          'enba_kea' => nil,
+          'enba_keap' => nil,
+          'enba_kga' => 137_043.0635298896,
+          'enba_kgap' => 'not enough data',
+          'enba_kha' => nil,
+          'enba_khap' => nil,
+          'enba_ksa' => nil,
+          'enba_ksap' => nil,
+          'enba_kxa' => 137_043.0635298896,
+          'enba_kxap' => nil,
+          'enba_cea' => nil,
+          'enba_ceap' => nil,
+          'enba_cga' => 28_779.043341276796,
+          'enba_cgap' => 'not enough data',
+          'enba_cha' => nil,
+          'enba_chap' => nil,
+          'enba_csa' => nil,
+          'enba_csap' => nil,
+          'enba_cxa' => 28_779.043341276796,
+          'enba_cxap' => nil,
+          'enba_pea' => nil,
+          'enba_peap' => nil,
+          'enba_pga' => 4111.291905896683,
+          'enba_pgap' => 'not enough data',
+          'enba_pha' => nil,
+          'enba_phap' => nil,
+          'enba_pxa' => 4111.291905896683,
+          'enba_pxap' => nil,
+          'enba_€ea' => nil,
+          'enba_€eap' => nil,
+          'enba_€ga' => 4111.291905896683,
+          'enba_€gap' => 'not enough data',
+          'enba_€ha' => nil,
+          'enba_€hap' => nil,
+          'enba_€sa' => nil,
+          'enba_€sap' => nil,
+          'enba_€xa' => 4111.291905896683,
+          'enba_€xap' => nil,
+          'enba_ratg' => 0.0,
+          'ihol_glyr' => 80.99353199999999,
+          'ihol_g£ly' => 80.99353199999999,
+          'ihol_fahl' => 1547,
+          'ihol_pnhl' => 203,
+          'ihol_gpfa' => 1.7451741435035544,
+          'ihol_epup' => 0.0,
+          'ihol_elyr' => 0.0,
+          'ihol_e£ly' => nil,
+          'ihol_pper' => 'Spring half term 2022',
+          'ihol_holp' => 0.08583469084374912,
+          'ihol_ahl£' => 296.61605099999997,
+          'ihol_esv€' => 1025.7307020892813,
+          'ihol_ratg' => 2.6,
+          'addp_pupn' => 203,
+          'addp_flra' => 1547.0,
+          'addp_sctp' => :primary,
+          'addp_stpn' => 'Primary',
+          'addp_name' => 'Acme Primary School 1',
+          'addp_area' => 'Acme Academy Trust',
+          'addp_urn' => 123_123,
+          'addp_ddays' => 1663.6166666666666,
+          'addp_etch' => nil,
+          'addp_etcp' => nil,
+          'addp_etpc' => nil,
+          'addp_etcd' => nil,
+          'addp_gtch' => false,
+          'addp_gtcp' => false,
+          'addp_gtpc' => nil,
+          'addp_gtcd' => nil,
+          'gsba_klyr' => 115_188.87762988935,
+          'gsba_kpyr' => 209_114.63420000012,
+          'gsba_£lyr' => 3455.666328896681,
+          'gsba_€lyr' => 3455.666328896681,
+          'gsba_£pyr' => 6273.439025999994,
+          'gsba_€pyr' => 6273.439025999994,
+          'gsba_co2y' => 24_189.664302276764,
+          'gsba_£exa' => 3300.4091909908816,
+          'gsba_€exa' => 3300.4091909908816,
+          'gsba_s£ex' => 155.25713790579948,
+          'gsba_s€ex' => 155.25713790579948,
+          'gsba_kpup' => 567.4328947285189,
+          'gsba_£pup' => 17.022986841855573,
+          'gsba_€pup' => 17.022986841855573,
+          'gsba_cpup' => 119.16090789298899,
+          'gsba_cfla' => 15.636499225776834,
+          'gsba_pfla' => 2.2337856036824055,
+          'gsba_€fla' => 2.2337856036824055,
+          'gsba_ddaj' => 1.124951418186212,
+          'gsba_ddly' => 1646.2979166666667,
+          'gsba_ddpy' => 1888.7145833333332,
+          'gsba_ddan' => -0.12835007936394124,
+          'gsba_kpya' => 213_548.60885858734,
+          'gsba_adpc' => -0.46059645040269104,
+          'gsba_nkpp' => 638.3344396503549,
+          'gsba_n£pp' => 19.15003318951065,
+          'gsba_n€pp' => 19.15003318951065,
+          'gsba_nkm2' => 83.76334275954883,
+          'gsba_n£m2' => 2.5129002827864655,
+          'gsba_n€m2' => 2.5129002827864655,
+          'gsba_pp%d' => -0.028830808584941074,
+          'gsba_ep%d' => 0.04704178449436038,
+          'gsba_ratg' => 6.39,
+          'hthe_oss£' => 104.71331100000002,
+          'hthe_oss€' => 104.71331100000002,
+          'hthe_htst' => '06:40',
+          'hthe_ratg' => 8.3,
+          'htsa_td1c' => 144.16749294913745,
+          'htsa_tc1c' => 1009.172450643962,
+          'htsa_ratg' => 5.0,
+          'hotw_ppyr' => 3.6364733399014786,
+          'hotw_eff' => 0.23460229740694694,
+          'hotw_gsav' => 216.70634500000006,
+          'hotw_esav' => -338.50528699999995,
+          'hotw_ratg' => 3.4,
+          'gswc_najk' => 6219.5515000000005,
+          'gswc_ajkw' => 5429.899116340053,
+          'gswc_difk' => 388.30225749373585,
+          'gswc_dif£' => 11.649067724812113,
+          'gswc_dif€' => 11.649067724812113,
+          'gswc_difp' => 0.06896253957462349,
+          'gswc_cppk' => 6018.928099999998,
+          'gswc_cppc' => 1263.974901,
+          'gswc_cpp£' => 180.56784299999998,
+          'gswc_cpp€' => 180.56784299999998,
+          'gswc_cper' => 'last school week (Sun 22-01-2023 to Sat 28-01-2023)',
+          'gswc_pppk' => 5630.625842506262,
+          'gswc_pppu' => 6219.5515000000005,
+          'gswc_ppp£' => 168.91877527518787,
+          'gswc_ppp€' => 168.91877527518787,
+          'gswc_pppc' => 1182.431426926315,
+          'gswc_pper' => 'previous school week (Sun 15-01-2023 to Sat 21-01-2023)',
+          'gswc_cptr' => false,
+          'gswc_cpfa' => 1547.0,
+          'gswc_ppfa' => 1547.0,
+          'gswc_fach' => false,
+          'gswc_cpnp' => 203,
+          'gswc_ppnp' => 203,
+          'gswc_pnch' => false,
+          'gswc_cp£k' => 0.030000000000000006,
+          'gswc_pp£k' => 0.03,
+          'gswc_cppp' => false,
+          'gswc_ratg' => 2.7,
+          'gphc_difk' => 2567.222546802109,
+          'gphc_dif£' => 77.01667640406329,
+          'gphc_dif€' => 77.01667640406329,
+          'gphc_difp' => 2.4943601583926345,
+          'gphc_cppk' => 3596.4333999999994,
+          'gphc_cppc' => 755.2510139999998,
+          'gphc_cpp£' => 107.893002,
+          'gphc_cpp€' => 107.893002,
+          'gphc_cper' => 'Xmas',
+          'gphc_pppk' => 1029.2108531978906,
+          'gphc_pppu' => 363.6693,
+          'gphc_ppp£' => 30.876325595936706,
+          'gphc_ppp€' => 30.876325595936706,
+          'gphc_pppc' => 216.13427917155695,
+          'gphc_pper' => 'Autumn half term',
+          'gphc_cptr' => false,
+          'gphc_cpfa' => 1547.0,
+          'gphc_ppfa' => 1547.0,
+          'gphc_fach' => false,
+          'gphc_cpnp' => 203,
+          'gphc_ppnp' => 203,
+          'gphc_pnch' => false,
+          'gphc_cp£k' => 0.030000000000000002,
+          'gphc_pp£k' => 0.02999999999999999,
+          'gphc_cppp' => false,
+          'gphc_uapk' => 363.6693,
+          'gphc_ratg' => 0.0,
+          'gpyc_difk' => -3484.0045211641204,
+          'gpyc_dif£' => -104.5201356349236,
+          'gpyc_dif€' => -104.5201356349236,
+          'gpyc_difp' => -0.49206059850480305,
+          'gpyc_cppk' => 3596.4333999999994,
+          'gpyc_cppc' => 755.2510139999998,
+          'gpyc_cpp£' => 107.893002,
+          'gpyc_cpp€' => 107.893002,
+          'gpyc_cper' => 'Xmas 2022/2023',
+          'gpyc_pppk' => 7080.43792116412,
+          'gpyc_pppu' => 6522.1486,
+          'gpyc_ppp£' => 212.4131376349236,
+          'gpyc_ppp€' => 212.4131376349236,
+          'gpyc_pppc' => 1486.891963444465,
+          'gpyc_pper' => 'Xmas 2021/2022',
+          'gpyc_cptr' => false,
+          'gpyc_cpfa' => 1547.0,
+          'gpyc_ppfa' => 1547.0,
+          'gpyc_fach' => false,
+          'gpyc_cpnp' => 203,
+          'gpyc_ppnp' => 203,
+          'gpyc_pnch' => false,
+          'gpyc_cp£k' => 0.030000000000000002,
+          'gpyc_pp£k' => 0.030000000000000002,
+          'gpyc_cppp' => false,
+          'gpyc_uapk' => 6522.1486,
+          'gpyc_ratg' => 10.0,
+          'gslt_ratg' => 10.0,
+          'opts_rmst' => 2.8412915417423226,
+          'opts_rmss' => 0.3584582576749295,
+          'opts_rmr2' => 0.030538933878712576,
+          'opts_avgt' => 5.65929203539823,
+          'opts_avhm' => '05:39',
+          'opts_sdst' => 1.9344913122493315,
+          'opts_ratg' => 8.9,
+          'shsd_wkwh' => 2341.4271582273245,
+          'shsd_wco2' => 491.69970322773827,
+          'shsd_w£__' => 70.24281474681976,
+          'shsd_w€__' => 70.24281474681976,
+          'shsd_wdys' => 5.0,
+          'shsd_wpan' => 0.028394778155501912,
+          'shsd_ratg' => 10.0,
+          'lug1_difk' => -297.7010563244852,
+          'lug1_dif£' => -8.931031689734555,
+          'lug1_dif€' => -8.931031689734555,
+          'lug1_difp' => -0.7117991270813318,
+          'lug1_cppk' => 120.5364,
+          'lug1_cppc' => 25.312644,
+          'lug1_cpp£' => 3.6160919999999996,
+          'lug1_cpp€' => 3.6160919999999996,
+          'lug1_cper' => 'Layer up power down day 11 November 2022',
+          'lug1_pppk' => 418.2374563244852,
+          'lug1_pppu' => 526.7040000000001,
+          'lug1_ppp£' => 12.547123689734555,
+          'lug1_ppp€' => 12.547123689734555,
+          'lug1_pppc' => 87.82986582814188,
+          'lug1_pper' => 'Layer up power down day 11 November 2022',
+          'lug1_cptr' => false,
+          'lug1_cpfa' => 1547.0,
+          'lug1_ppfa' => 1547.0,
+          'lug1_fach' => false,
+          'lug1_cpnp' => 203,
+          'lug1_ppnp' => 203,
+          'lug1_pnch' => false,
+          'lug1_cp£k' => 0.029999999999999995,
+          'lug1_pp£k' => 0.029999999999999995,
+          'lug1_cppp' => false,
+          'lug1_uapk' => 526.7040000000001,
+          'lug1_ratg' => 10.0,
+          'a22g_difk' => -19_528.49596725549,
+          'a22g_dif£' => -585.8548790176653,
+          'a22g_dif€' => -585.8548790176653,
+          'a22g_difp' => -0.37732009066988376,
+          'a22g_cppk' => 32_227.28500000001,
+          'a22g_cppc' => 6767.7298500000015,
+          'a22g_cpp£' => 966.81855,
+          'a22g_cpp€' => 966.81855,
+          'a22g_cper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          'a22g_pppk' => 51_755.7809672555,
+          'a22g_pppu' => 58_144.8113,
+          'a22g_ppp£' => 1552.6734290176653,
+          'a22g_ppp€' => 1552.6734290176653,
+          'a22g_pppc' => 10_868.714003123656,
+          'a22g_pper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          'a22g_cptr' => false,
+          'a22g_cpfa' => 1547.0,
+          'a22g_ppfa' => 1547.0,
+          'a22g_fach' => false,
+          'a22g_cpnp' => 203,
+          'a22g_ppnp' => 203,
+          'a22g_pnch' => false,
+          'a22g_cp£k' => 0.02999999999999999,
+          'a22g_pp£k' => 0.030000000000000006,
+          'a22g_cppp' => false,
+          'a22g_uapk' => 58_144.81129999999,
+          'a22g_ratg' => 10.0,
+          's22g_difk' => -19_528.49596725549,
+          's22g_dif£' => -585.8548790176653,
+          's22g_dif€' => -585.8548790176653,
+          's22g_difp' => -0.37732009066988376,
+          's22g_cppk' => 32_227.28500000001,
+          's22g_cppc' => 6767.7298500000015,
+          's22g_cpp£' => 966.81855,
+          's22g_cpp€' => 966.81855,
+          's22g_cper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          's22g_pppk' => 51_755.7809672555,
+          's22g_pppu' => 58_144.8113,
+          's22g_ppp£' => 1552.6734290176653,
+          's22g_ppp€' => 1552.6734290176653,
+          's22g_pppc' => 10_868.714003123656,
+          's22g_pper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          's22g_cptr' => false,
+          's22g_cpfa' => 1547.0,
+          's22g_ppfa' => 1547.0,
+          's22g_fach' => false,
+          's22g_cpnp' => 203,
+          's22g_ppnp' => 203,
+          's22g_pnch' => false,
+          's22g_cp£k' => 0.02999999999999999,
+          's22g_pp£k' => 0.030000000000000006,
+          's22g_cppp' => false,
+          's22g_uapk' => 58_144.81129999999,
+          's22g_ratg' => 10.0
+        },
+        629 => {
+          'enba_klyr' => 268_290.25806213793,
+          'enba_£lyr' => 8048.707741864137,
+          'enba_€lyr' => 8048.707741864137,
+          'enba_co2y' => 56_340.95419304895,
+          'enba_co2t' => 56.34095419304895,
+          'enba_co2e' => nil,
+          'enba_co2g' => 56_340.95419304895,
+          'enba_co2h' => nil,
+          'enba_co2s' => nil,
+          'enba_kpup' => 596.2005734714177,
+          'enba_£pup' => 17.886017204142526,
+          'enba_cpup' => 125.20212042899767,
+          'enba_kfla' => 89.22190158368404,
+          'enba_£fla' => 2.676657047510521,
+          'enba_cfla' => 18.736599332573643,
+          'enba_pp%d' => 0.2918485726816435,
+          'enba_sact' => Date.parse('Fri, 21 Oct 2022'),
+          'enba_solr' => '',
+          'enba_ke0' => nil,
+          'enba_ken' => nil,
+          'enba_kenp' => nil,
+          'enba_kg0' => 268_290.25806213793,
+          'enba_kgn' => 338_621.6341999999,
+          'enba_kgnp' => -0.2076990039458678,
+          'enba_kh0' => nil,
+          'enba_khn' => nil,
+          'enba_khnp' => nil,
+          'enba_ks0' => nil,
+          'enba_ksn' => nil,
+          'enba_ksnp' => nil,
+          'enba_kx0' => 268_290.25806213793,
+          'enba_kxn' => 338_621.6341999999,
+          'enba_kxnp' => -0.2076990039458678,
+          'enba_ce0' => nil,
+          'enba_cen' => nil,
+          'enba_cenp' => nil,
+          'enba_cg0' => 56_340.95419304895,
+          'enba_cgn' => 71_110.54318200007,
+          'enba_cgnp' => -0.20769900394586896,
+          'enba_ch0' => nil,
+          'enba_chn' => nil,
+          'enba_chnp' => nil,
+          'enba_cs0' => nil,
+          'enba_csn' => nil,
+          'enba_csnp' => nil,
+          'enba_cx0' => 56_340.95419304895,
+          'enba_cxn' => 71_110.54318200007,
+          'enba_cxnp' => -0.20769900394586896,
+          'enba_pe0' => nil,
+          'enba_pen' => nil,
+          'enba_penp' => nil,
+          'enba_pg0' => 8048.707741864137,
+          'enba_pgn' => 10_158.649026000003,
+          'enba_pgnp' => -0.20769900394586835,
+          'enba_ph0' => nil,
+          'enba_phn' => nil,
+          'enba_phnp' => nil,
+          'enba_px0' => 8048.707741864137,
+          'enba_pxn' => 10_158.649026000003,
+          'enba_pxnp' => -0.20769900394586835,
+          'enba_€e0' => nil,
+          'enba_€en' => nil,
+          'enba_€enp' => nil,
+          'enba_€g0' => 8048.707741864137,
+          'enba_€gn' => 10_158.649026000003,
+          'enba_€gnp' => -0.20769900394586835,
+          'enba_€h0' => nil,
+          'enba_€hn' => nil,
+          'enba_€hnp' => nil,
+          'enba_€s0' => nil,
+          'enba_€sn' => nil,
+          'enba_€snp' => nil,
+          'enba_€x0' => 8048.707741864137,
+          'enba_€xn' => 10_158.649026000003,
+          'enba_€xnp' => -0.20769900394586835,
+          'enba_kea' => nil,
+          'enba_keap' => nil,
+          'enba_kga' => 297_577.7828621378,
+          'enba_kgap' => 'not enough data',
+          'enba_kha' => nil,
+          'enba_khap' => nil,
+          'enba_ksa' => nil,
+          'enba_ksap' => nil,
+          'enba_kxa' => 297_577.7828621378,
+          'enba_kxap' => nil,
+          'enba_cea' => nil,
+          'enba_ceap' => nil,
+          'enba_cga' => 62_491.33440104901,
+          'enba_cgap' => 'not enough data',
+          'enba_cha' => nil,
+          'enba_chap' => nil,
+          'enba_csa' => nil,
+          'enba_csap' => nil,
+          'enba_cxa' => 62_491.33440104901,
+          'enba_cxap' => nil,
+          'enba_pea' => nil,
+          'enba_peap' => nil,
+          'enba_pga' => 8927.333485864141,
+          'enba_pgap' => 'not enough data',
+          'enba_pha' => nil,
+          'enba_phap' => nil,
+          'enba_pxa' => 8927.333485864141,
+          'enba_pxap' => nil,
+          'enba_€ea' => nil,
+          'enba_€eap' => nil,
+          'enba_€ga' => 8927.333485864141,
+          'enba_€gap' => 'not enough data',
+          'enba_€ha' => nil,
+          'enba_€hap' => nil,
+          'enba_€sa' => nil,
+          'enba_€sap' => nil,
+          'enba_€xa' => 8927.333485864141,
+          'enba_€xap' => nil,
+          'enba_ratg' => 0.0,
+          'ihol_glyr' => 406.15747799999997,
+          'ihol_g£ly' => 406.15747799999997,
+          'ihol_fahl' => 3007,
+          'ihol_pnhl' => 450,
+          'ihol_gpfa' => 4.502355370801464,
+          'ihol_epup' => 0.0,
+          'ihol_elyr' => 0.0,
+          'ihol_e£ly' => nil,
+          'ihol_pper' => 'Spring half term 2022',
+          'ihol_holp' => 0.21875650097741126,
+          'ihol_ahl£' => 1760.7071429999992,
+          'ihol_esv€' => 4028.6261021113023,
+          'ihol_ratg' => 0.0,
+          'addp_pupn' => 450,
+          'addp_flra' => 3007.0,
+          'addp_sctp' => :primary,
+          'addp_stpn' => 'Primary',
+          'addp_name' => 'Acme Primary School 2',
+          'addp_area' => 'Acme Academy Trust',
+          'addp_urn' => 123_345,
+          'addp_ddays' => 1663.6166666666666,
+          'addp_etch' => nil,
+          'addp_etcp' => nil,
+          'addp_etpc' => nil,
+          'addp_etcd' => nil,
+          'addp_gtch' => false,
+          'addp_gtcp' => false,
+          'addp_gtpc' => nil,
+          'addp_gtcd' => nil,
+          'gsba_klyr' => 268_290.25806213793,
+          'gsba_kpyr' => 338_621.6341999999,
+          'gsba_£lyr' => 8048.707741864137,
+          'gsba_€lyr' => 8048.707741864137,
+          'gsba_£pyr' => 10_158.649026000003,
+          'gsba_€pyr' => 10_158.649026000003,
+          'gsba_co2y' => 56_340.95419304895,
+          'gsba_£exa' => 6415.210366715952,
+          'gsba_€exa' => 6415.210366715952,
+          'gsba_s£ex' => 1633.4973751481848,
+          'gsba_s€ex' => 1633.4973751481848,
+          'gsba_kpup' => 596.2005734714177,
+          'gsba_£pup' => 17.886017204142526,
+          'gsba_€pup' => 17.886017204142526,
+          'gsba_cpup' => 125.20212042899767,
+          'gsba_cfla' => 18.736599332573643,
+          'gsba_pfla' => 2.676657047510521,
+          'gsba_€fla' => 2.676657047510521,
+          'gsba_ddaj' => 1.124951418186212,
+          'gsba_ddly' => 1646.2979166666667,
+          'gsba_ddpy' => 1888.7145833333332,
+          'gsba_ddan' => -0.12835007936394124,
+          'gsba_kpya' => 341_532.3764669895,
+          'gsba_adpc' => -0.2144514647850106,
+          'gsba_nkpp' => 670.6966806501041,
+          'gsba_n£pp' => 20.12090041950312,
+          'gsba_n€pp' => 20.12090041950312,
+          'gsba_nkm2' => 100.37030471983599,
+          'gsba_n£m2' => 3.0111091415950795,
+          'gsba_n€m2' => 3.0111091415950795,
+          'gsba_pp%d' => 0.16371367791114194,
+          'gsba_ep%d' => 0.25462880899795004,
+          'gsba_ratg' => 0.0,
+          'hthe_oss£' => 375.29613600000016,
+          'hthe_oss€' => 375.29613600000016,
+          'hthe_htst' => '02:45',
+          'hthe_ratg' => 5.0,
+          'htsa_td1c' => 80.01466766613277,
+          'htsa_tc1c' => 560.1026736629294,
+          'htsa_ratg' => 5.0,
+          'hotw_ppyr' => 8.869078091249998,
+          'hotw_eff' => 0.09619094467571224,
+          'hotw_gsav' => 2786.312351062499,
+          'hotw_esav' => 1630.816391062499,
+          'hotw_ratg' => 0.8,
+          'gsoo_sdop' => 0.19946920284395286,
+          'gsoo_sdcp' => 0.4075116392424616,
+          'gsoo_holp' => 0.21875650097741126,
+          'gsoo_wkep' => 0.17426265693617415,
+          'gsoo_comp' => 0.0,
+          'gsoo_ahl£' => 1760.7071429999992,
+          'gsoo_awk£' => 1402.589196,
+          'gsoo_com£' => 0.0,
+          'gsoo_aoo£' => 6443.238424670543,
+          'gsoo_ahl€' => 1760.7071429999992,
+          'gsoo_awk€' => 1402.589196,
+          'gsoo_com€' => 0.0,
+          'gsoo_aoo€' => 6443.238424670543,
+          'gsoo_esv€' => 4028.6261021113023,
+          'gsoo_ratg' => 0.0,
+          'hwia_ratg' => 9,
+          'httc_r2' => 0.17710417669937942,
+          'httc_sav€' => 473.9148060183556,
+          'httc_ratg' => 1,
+          'gswe_ratg' => 0.0,
+          'gswc_najk' => 11_075.2998,
+          'gswc_ajkw' => 10_830.737272604694,
+          'gswc_difk' => -270.30697652378694,
+          'gswc_dif£' => -8.109209295713583,
+          'gswc_dif€' => -8.109209295713583,
+          'gswc_difp' => -0.024985446063124694,
+          'gswc_cppk' => 10_548.2702,
+          'gswc_cppc' => 2215.1367419999997,
+          'gswc_cpp£' => 316.44810600000005,
+          'gswc_cpp€' => 316.44810600000005,
+          'gswc_cper' => 'last school week (Sun 22-01-2023 to Sat 28-01-2023)',
+          'gswc_pppk' => 10_818.577176523788,
+          'gswc_pppu' => 11_075.2998,
+          'gswc_ppp£' => 324.55731529571364,
+          'gswc_ppp€' => 324.55731529571364,
+          'gswc_pppc' => 2271.9012070699955,
+          'gswc_pper' => 'previous school week (Sun 15-01-2023 to Sat 21-01-2023)',
+          'gswc_cptr' => false,
+          'gswc_cpfa' => 3007.0,
+          'gswc_ppfa' => 3007.0,
+          'gswc_fach' => false,
+          'gswc_cpnp' => 450,
+          'gswc_ppnp' => 450,
+          'gswc_pnch' => false,
+          'gswc_cp£k' => 0.030000000000000002,
+          'gswc_pp£k' => 0.03,
+          'gswc_cppp' => false,
+          'gswc_ratg' => 10.0,
+          'gphc_difk' => 10_949.952554450407,
+          'gphc_dif£' => 328.49857663351236,
+          'gphc_dif€' => 328.49857663351236,
+          'gphc_difp' => 1.8861220156297185,
+          'gphc_cppk' => 16_755.490299999998,
+          'gphc_cppc' => 3518.652963,
+          'gphc_cpp£' => 502.664709,
+          'gphc_cpp€' => 502.664709,
+          'gphc_cper' => 'Xmas',
+          'gphc_pppk' => 5805.537745549591,
+          'gphc_pppu' => 3141.9899000000005,
+          'gphc_ppp£' => 174.1661323664877,
+          'gphc_ppp€' => 174.1661323664877,
+          'gphc_pppc' => 1219.1629265654137,
+          'gphc_pper' => 'Autumn half term',
+          'gphc_cptr' => false,
+          'gphc_cpfa' => 3007.0,
+          'gphc_ppfa' => 3007.0,
+          'gphc_fach' => false,
+          'gphc_cpnp' => 450,
+          'gphc_ppnp' => 450,
+          'gphc_pnch' => false,
+          'gphc_cp£k' => 0.030000000000000006,
+          'gphc_pp£k' => 0.029999999999999992,
+          'gphc_cppp' => false,
+          'gphc_uapk' => 3141.9899000000005,
+          'gphc_ratg' => 0.0,
+          'gpyc_difk' => -1836.4714425878847,
+          'gpyc_dif£' => -55.094143277636476,
+          'gpyc_dif€' => -55.094143277636476,
+          'gpyc_difp' => -0.09877771200341656,
+          'gpyc_cppk' => 16_755.490299999998,
+          'gpyc_cppc' => 3518.652963,
+          'gpyc_cpp£' => 502.664709,
+          'gpyc_cpp€' => 502.664709,
+          'gpyc_cper' => 'Xmas 2022/2023',
+          'gpyc_pppk' => 18_591.961742587882,
+          'gpyc_pppu' => 22_712.1188,
+          'gpyc_ppp£' => 557.7588522776365,
+          'gpyc_ppp€' => 557.7588522776365,
+          'gpyc_pppc' => 3904.3119659434547,
+          'gpyc_pper' => 'Xmas 2021/2022',
+          'gpyc_cptr' => false,
+          'gpyc_cpfa' => 3007.0,
+          'gpyc_ppfa' => 3007.0,
+          'gpyc_fach' => false,
+          'gpyc_cpnp' => 450,
+          'gpyc_ppnp' => 450,
+          'gpyc_pnch' => false,
+          'gpyc_cp£k' => 0.030000000000000006,
+          'gpyc_pp£k' => 0.030000000000000002,
+          'gpyc_cppp' => false,
+          'gpyc_uapk' => 22_712.118799999997,
+          'gpyc_ratg' => 8.3,
+          'gslt_ratg' => 9.2,
+          'opts_rmst' => 4.509564074357204,
+          'opts_rmss' => 0.11594307609121246,
+          'opts_rmr2' => 0.010002588486837872,
+          'opts_avgt' => 3.172932330827068,
+          'opts_avhm' => '03:10',
+          'opts_sdst' => 3.4956927570149534,
+          'opts_ratg' => 0.6,
+          'shsd_wkwh' => 4411.459374914168,
+          'shsd_wco2' => 926.406468731975,
+          'shsd_w£__' => 132.343781247425,
+          'shsd_w€__' => 132.343781247425,
+          'shsd_wdys' => 5.0,
+          'shsd_wpan' => 0.037642203876895534,
+          'shsd_ratg' => 9.2,
+          'lug1_difk' => -493.0521230319683,
+          'lug1_dif£' => -14.791563690959059,
+          'lug1_dif€' => -14.791563690959059,
+          'lug1_difp' => -0.5738339248925467,
+          'lug1_cppk' => 366.1723000000001,
+          'lug1_cppc' => 76.896183,
+          'lug1_cpp£' => 10.985168999999999,
+          'lug1_cpp€' => 10.985168999999999,
+          'lug1_cper' => 'Layer up power down day 11 November 2022',
+          'lug1_pppk' => 859.2244230319684,
+          'lug1_pppu' => 897.5110999999997,
+          'lug1_ppp£' => 25.776732690959058,
+          'lug1_ppp€' => 25.776732690959058,
+          'lug1_pppc' => 180.4371288367134,
+          'lug1_pper' => 'Layer up power down day 11 November 2022',
+          'lug1_cptr' => false,
+          'lug1_cpfa' => 3007.0,
+          'lug1_ppfa' => 3007.0,
+          'lug1_fach' => false,
+          'lug1_cpnp' => 450,
+          'lug1_ppnp' => 450,
+          'lug1_pnch' => false,
+          'lug1_cp£k' => 0.02999999999999999,
+          'lug1_pp£k' => 0.030000000000000006,
+          'lug1_cppp' => false,
+          'lug1_uapk' => 897.5110999999997,
+          'lug1_ratg' => 10.0,
+          'a22g_difk' => -20_361.04555800622,
+          'a22g_dif£' => -610.8313667401871,
+          'a22g_dif€' => -610.8313667401871,
+          'a22g_difp' => -0.23788911285025313,
+          'a22g_cppk' => 65_229.443699999996,
+          'a22g_cppc' => 13_698.183177,
+          'a22g_cpp£' => 1956.8833109999996,
+          'a22g_cpp€' => 1956.8833109999996,
+          'a22g_cper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          'a22g_pppk' => 85_590.48925800622,
+          'a22g_pppu' => 88_235.4562,
+          'a22g_ppp£' => 2567.7146777401867,
+          'a22g_ppp€' => 2567.7146777401867,
+          'a22g_pppc' => 17_974.002744181307,
+          'a22g_pper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          'a22g_cptr' => false,
+          'a22g_cpfa' => 3007.0,
+          'a22g_ppfa' => 3007.0,
+          'a22g_fach' => false,
+          'a22g_cpnp' => 450,
+          'a22g_ppnp' => 450,
+          'a22g_pnch' => false,
+          'a22g_cp£k' => 0.029999999999999995,
+          'a22g_pp£k' => 0.030000000000000002,
+          'a22g_cppp' => false,
+          'a22g_uapk' => 88_235.45619999999,
+          'a22g_ratg' => 10.0,
+          's22g_difk' => -20_361.04555800622,
+          's22g_dif£' => -610.8313667401871,
+          's22g_dif€' => -610.8313667401871,
+          's22g_difp' => -0.23788911285025313,
+          's22g_cppk' => 65_229.443699999996,
+          's22g_cppc' => 13_698.183177,
+          's22g_cpp£' => 1956.8833109999996,
+          's22g_cpp€' => 1956.8833109999996,
+          's22g_cper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          's22g_pppk' => 85_590.48925800622,
+          's22g_pppu' => 88_235.4562,
+          's22g_ppp£' => 2567.7146777401867,
+          's22g_ppp€' => 2567.7146777401867,
+          's22g_pppc' => 17_974.002744181307,
+          's22g_pper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          's22g_cptr' => false,
+          's22g_cpfa' => 3007.0,
+          's22g_ppfa' => 3007.0,
+          's22g_fach' => false,
+          's22g_cpnp' => 450,
+          's22g_ppnp' => 450,
+          's22g_pnch' => false,
+          's22g_cp£k' => 0.029999999999999995,
+          's22g_pp£k' => 0.030000000000000002,
+          's22g_cppp' => false,
+          's22g_uapk' => 88_235.45619999999,
+          's22g_ratg' => 10.0
+        },
+        634 => {
+          'enba_klyr' => nil,
+          'enba_£lyr' => nil,
+          'enba_€lyr' => nil,
+          'enba_co2y' => nil,
+          'enba_co2t' => nil,
+          'enba_co2e' => 82_218.20810000003,
+          'enba_co2g' => nil,
+          'enba_co2h' => nil,
+          'enba_co2s' => nil,
+          'enba_kpup' => 361.7098785425103,
+          'enba_£pup' => 54.25648178137652,
+          'enba_cpup' => 66.57344785425104,
+          'enba_kfla' => 44.936294135398875,
+          'enba_£fla' => 6.740444120309829,
+          'enba_cfla' => 8.270617452972541,
+          'enba_pp%d' => 0.11295347243849324,
+          'enba_sact' => Date.parse('Mon, 31 Oct 2022'),
+          'enba_solr' => '',
+          'enba_ke0' => 446_711.70000000024,
+          'enba_ken' => 456_813.90000000043,
+          'enba_kenp' => -0.02211447593867038,
+          'enba_kg0' => nil,
+          'enba_kgn' => nil,
+          'enba_kgnp' => 'not enough data',
+          'enba_kh0' => nil,
+          'enba_khn' => nil,
+          'enba_khnp' => nil,
+          'enba_ks0' => nil,
+          'enba_ksn' => nil,
+          'enba_ksnp' => nil,
+          'enba_kx0' => nil,
+          'enba_kxn' => nil,
+          'enba_kxnp' => nil,
+          'enba_ce0' => 82_218.20810000003,
+          'enba_cen' => 87_990.72848000002,
+          'enba_cenp' => -0.06560373439017567,
+          'enba_cg0' => nil,
+          'enba_cgn' => nil,
+          'enba_cgnp' => 'not enough data',
+          'enba_ch0' => nil,
+          'enba_chn' => nil,
+          'enba_chnp' => nil,
+          'enba_cs0' => nil,
+          'enba_csn' => nil,
+          'enba_csnp' => nil,
+          'enba_cx0' => nil,
+          'enba_cxn' => nil,
+          'enba_cxnp' => nil,
+          'enba_pe0' => 67_006.755,
+          'enba_pen' => 68_522.08500000004,
+          'enba_penp' => -0.02211447593867043,
+          'enba_pg0' => nil,
+          'enba_pgn' => nil,
+          'enba_pgnp' => 'not enough data',
+          'enba_ph0' => nil,
+          'enba_phn' => nil,
+          'enba_phnp' => nil,
+          'enba_px0' => nil,
+          'enba_pxn' => nil,
+          'enba_pxnp' => nil,
+          'enba_€e0' => 67_006.755,
+          'enba_€en' => 68_522.08500000004,
+          'enba_€enp' => -0.02211447593867043,
+          'enba_€g0' => nil,
+          'enba_€gn' => nil,
+          'enba_€gnp' => 'not enough data',
+          'enba_€h0' => nil,
+          'enba_€hn' => nil,
+          'enba_€hnp' => nil,
+          'enba_€s0' => nil,
+          'enba_€sn' => nil,
+          'enba_€snp' => nil,
+          'enba_€x0' => nil,
+          'enba_€xn' => nil,
+          'enba_€xnp' => nil,
+          'enba_kea' => 446_836.10000000044,
+          'enba_keap' => 'not enough data',
+          'enba_kga' => nil,
+          'enba_kgap' => 'not enough data',
+          'enba_kha' => nil,
+          'enba_khap' => nil,
+          'enba_ksa' => nil,
+          'enba_ksap' => nil,
+          'enba_kxa' => nil,
+          'enba_kxap' => nil,
+          'enba_cea' => 85_825.67167999991,
+          'enba_ceap' => 'not enough data',
+          'enba_cga' => nil,
+          'enba_cgap' => 'not enough data',
+          'enba_cha' => nil,
+          'enba_chap' => nil,
+          'enba_csa' => nil,
+          'enba_csap' => nil,
+          'enba_cxa' => nil,
+          'enba_cxap' => nil,
+          'enba_pea' => 67_025.41499999996,
+          'enba_peap' => 'not enough data',
+          'enba_pga' => nil,
+          'enba_pgap' => 'not enough data',
+          'enba_pha' => nil,
+          'enba_phap' => nil,
+          'enba_pxa' => nil,
+          'enba_pxap' => nil,
+          'enba_€ea' => 67_025.41499999996,
+          'enba_€eap' => 'not enough data',
+          'enba_€ga' => nil,
+          'enba_€gap' => 'not enough data',
+          'enba_€ha' => nil,
+          'enba_€hap' => nil,
+          'enba_€sa' => nil,
+          'enba_€sap' => nil,
+          'enba_€xa' => nil,
+          'enba_€xap' => nil,
+          'enba_ratg' => 1.7,
+          'addp_pupn' => 1235,
+          'addp_flra' => 9941.0,
+          'addp_sctp' => :secondary,
+          'addp_stpn' => 'Secondary',
+          'addp_name' => 'Acme Secondary School 3',
+          'addp_area' => 'Acme Academy Trust',
+          'addp_urn' => 123_678,
+          'addp_ddays' => 1663.6166666666666,
+          'addp_etch' => false,
+          'addp_etcp' => false,
+          'addp_etpc' => nil,
+          'addp_etcd' => nil,
+          'addp_gtch' => false,
+          'addp_gtcp' => nil,
+          'addp_gtpc' => nil,
+          'addp_gtcd' => nil,
+          'elbc_blly' => 24.570753424657536,
+          'elbc_bllw' => 34.021428571428565,
+          'elbc_blch' => 9.45067514677103,
+          'elbc_anc€' => 12_418.187142857128,
+          'elbc_bspc' => 0.3846310686300313,
+          'elbc_€prk' => 0.14999999999999997,
+          'elbc_abkp' => 0.48108352519846037,
+          'elbc_ratg' => 0.0,
+          'elba_klyr' => 446_711.70000000024,
+          'elba_£lyr' => 67_006.755,
+          'elba_€lyr' => 67_006.755,
+          'elba_kpyr' => 456_813.90000000043,
+          'elba_£pyr' => 68_522.08500000004,
+          'elba_€pyr' => 68_522.08500000004,
+          'elba_co2y' => 82_218.20810000003,
+          'elba_£esav' => 24_862.380000000026,
+          'elba_€esav' => 24_862.380000000026,
+          'elba_kpup' => 361.7098785425103,
+          'elba_£pup' => 54.25648178137652,
+          'elba_€pup' => 54.25648178137652,
+          'elba_cpup' => 66.57344785425104,
+          'elba_c£m2' => 8.270617452972541,
+          'elba_pp%d' => 0.1129534724384933,
+          'elba_ep%d' => 0.5899335320549904,
+          'elba_ratg' => 1.74,
+          'elbb_lykw' => 24.570753424657536,
+          'elbb_lygb' => 32_285.969999999998,
+          'elbb_svex' => 8633.970000000001,
+          'elbb_blpp' => 0.019895346902556708,
+          'elbb_€prk' => 0.14999999999999997,
+          'elbb_abkp' => 0.48108352519846037,
+          'elbb_ratg' => 6.71,
+          'eloo_sdop' => 0.38783373407651206,
+          'eloo_sdcp' => 0.3563337009828338,
+          'eloo_holp' => 0.1501621739479848,
+          'eloo_wkep' => 0.10567039099266931,
+          'eloo_comp' => 0.0,
+          'eloo_ahl£' => 10_061.879999999996,
+          'eloo_awk£' => 7080.63,
+          'eloo_com£' => 0.0,
+          'eloo_aoo£' => 41_019.274999999994,
+          'eloo_ahl€' => 10_061.879999999996,
+          'eloo_awk€' => 7080.63,
+          'eloo_com€' => 0.0,
+          'eloo_aoo€' => 41_019.274999999994,
+          'eloo_esv€' => 17_566.910750000003,
+          'eloo_ratg' => 1.3,
+          'emtc_mets' => 3,
+          'emtc_sav£' => 2190.0,
+          'emtc_ratg' => 0.0,
+          'dtaf_sav£' => 0.0,
+          'dtaf_ratg' => 10.0,
+          'eswc_difk' => 316.1999999999989,
+          'eswc_dif£' => 47.429999999999836,
+          'eswc_dif€' => 47.429999999999836,
+          'eswc_difp' => 0.02384759261493898,
+          'eswc_cppk' => 13_575.4,
+          'eswc_cppc' => 3150.3226999999997,
+          'eswc_cpp£' => 2036.31,
+          'eswc_cpp€' => 2036.31,
+          'eswc_cper' => 'last school week (Sun 22-01-2023 to Sat 28-01-2023)',
+          'eswc_pppk' => 13_259.2,
+          'eswc_pppu' => 13_259.2,
+          'eswc_ppp£' => 1988.88,
+          'eswc_ppp€' => 1988.88,
+          'eswc_pppc' => 2615.9201,
+          'eswc_pper' => 'previous school week (Sun 15-01-2023 to Sat 21-01-2023)',
+          'eswc_cptr' => false,
+          'eswc_cpfa' => 9941.0,
+          'eswc_ppfa' => 9941.0,
+          'eswc_fach' => false,
+          'eswc_cpnp' => 1235,
+          'eswc_ppnp' => 1235,
+          'eswc_pnch' => false,
+          'eswc_cp£k' => 0.15,
+          'eswc_pp£k' => 0.15,
+          'eswc_cppp' => false,
+          'eswc_ratg' => 3.8,
+          'ephc_difk' => -1470.1399999999994,
+          'ephc_dif£' => -220.52099999999996,
+          'ephc_dif€' => -220.52099999999996,
+          'ephc_difp' => -0.11306994902338399,
+          'ephc_cppk' => 11_531.9,
+          'ephc_cppc' => 1173.8665,
+          'ephc_cpp£' => 1729.785,
+          'ephc_cpp€' => 1729.785,
+          'ephc_cper' => 'Xmas',
+          'ephc_pppk' => 13_002.039999999999,
+          'ephc_pppu' => 7402.599999999999,
+          'ephc_ppp£' => 1950.306,
+          'ephc_ppp€' => 1950.306,
+          'ephc_pppc' => 2097.59186,
+          'ephc_pper' => 'Autumn half term',
+          'ephc_cptr' => false,
+          'ephc_cpfa' => 9941.0,
+          'ephc_ppfa' => 9941.0,
+          'ephc_fach' => false,
+          'ephc_cpnp' => 1235,
+          'ephc_ppnp' => 1235,
+          'ephc_pnch' => false,
+          'ephc_cp£k' => 0.15000000000000002,
+          'ephc_pp£k' => 0.15000000000000002,
+          'ephc_cppp' => false,
+          'ephc_ratg' => 10.0,
+          'epyc_difk' => -388.5600000000013,
+          'epyc_dif£' => -58.28399999999988,
+          'epyc_dif€' => -58.28399999999988,
+          'epyc_difp' => -0.03259605753469256,
+          'epyc_cppk' => 11_531.9,
+          'epyc_cppc' => 1173.8665,
+          'epyc_cpp£' => 1729.785,
+          'epyc_cpp€' => 1729.785,
+          'epyc_cper' => 'Xmas 2022/2023',
+          'epyc_pppk' => 11_920.460000000001,
+          'epyc_pppu' => 11_057.1,
+          'epyc_ppp£' => 1788.069,
+          'epyc_ppp€' => 1788.069,
+          'epyc_pppc' => 1633.67335,
+          'epyc_pper' => 'Xmas 2021/2022',
+          'epyc_cptr' => false,
+          'epyc_cpfa' => 9941.0,
+          'epyc_ppfa' => 9941.0,
+          'epyc_fach' => false,
+          'epyc_cpnp' => 1235,
+          'epyc_ppnp' => 1235,
+          'epyc_pnch' => false,
+          'epyc_cp£k' => 0.15000000000000002,
+          'epyc_pp£k' => 0.15,
+          'epyc_cppp' => false,
+          'epyc_ratg' => 6.2,
+          'epkb_kwsc' => 151.67972508591063,
+          'epkb_kwpp' => 0.12281759116268068,
+          'epkb_kwfa' => 0.015257994677186464,
+          'epkb_kwex' => 99.41,
+          'epkb_tex£' => 8806.169249999999,
+          'epkb_ratg' => 4.7,
+          'sole_opvk' => 414.0,
+          'sole_opvy' => 4.783457104878385,
+          'sole_opvp' => 0.38790836461322636,
+          'sole_opv€' => 36_983.72012567629,
+          'sole_ratg' => 5.0,
+          'ellt_ratg' => 6.9,
+          'free_ann£' => 2016.9493031358854,
+          'free_hol£' => 225.75999999999968,
+          'free_kwrd' => -1.5349690282617097,
+          'free_rrat' => 10.0,
+          'free_trat' => 0.0,
+          'free_ratg' => 0.0,
+          'sblv_wtbl' => 33.56333333333333,
+          'sblv_smbl' => 19.888333333333332,
+          'sblv_sblp' => 0.6875890387999666,
+          'sblv_ckwh' => 44_366.80000000001,
+          'sblv_cgbp' => 6655.02,
+          'sblv_c€bp' => 6655.02,
+          'sblv_€prk' => 0.14999999999999997,
+          'sblv_abkp' => 0.48108352519846037,
+          'sblv_ratg' => 0.0,
+          'iblv_mxbk' => 26.752205882352943,
+          'iblv_mxbd' => 'Sunday',
+          'iblv_mnbk' => 24.84625,
+          'iblv_mnbd' => 'Friday',
+          'iblv_sblp' => 0.07671000180521975,
+          'iblv_ckwh' => 8653.205468440563,
+          'iblv_cgbp' => 1297.9808202660843,
+          'iblv_c€bp' => 1297.9808202660843,
+          'iblv_€prk' => 0.14999999999999997,
+          'iblv_abkp' => 0.48108352519846037,
+          'iblv_ratg' => 10.0,
+          'lue1_difk' => -15.299999999999272,
+          'lue1_dif£' => -2.295000000000016,
+          'lue1_dif€' => -2.295000000000016,
+          'lue1_difp' => -0.007922945471492556,
+          'lue1_cppk' => 1915.8000000000004,
+          'lue1_cppc' => 209.4307,
+          'lue1_cpp£' => 287.37,
+          'lue1_cpp€' => 287.37,
+          'lue1_cper' => 'Layer up power down day 11 November 2022',
+          'lue1_pppk' => 1931.0999999999997,
+          'lue1_pppu' => 1931.0999999999997,
+          'lue1_ppp£' => 289.665,
+          'lue1_ppp€' => 289.665,
+          'lue1_pppc' => 247.5078,
+          'lue1_pper' => 'Layer up power down day 11 November 2022',
+          'lue1_cptr' => false,
+          'lue1_cpfa' => 9941.0,
+          'lue1_ppfa' => 9941.0,
+          'lue1_fach' => false,
+          'lue1_cpnp' => 1235,
+          'lue1_ppnp' => 1235,
+          'lue1_pnch' => false,
+          'lue1_cp£k' => 0.14999999999999997,
+          'lue1_pp£k' => 0.15000000000000002,
+          'lue1_cppp' => false,
+          'lue1_ratg' => 5.4,
+          'a22e_difk' => -5857.8000000000175,
+          'a22e_dif£' => -878.6700000000055,
+          'a22e_dif€' => -878.6700000000055,
+          'a22e_difp' => -0.03753024850383239,
+          'a22e_cppk' => 150_224.3,
+          'a22e_cppc' => 29_528.102099999996,
+          'a22e_cpp£' => 22_533.644999999997,
+          'a22e_cpp€' => 22_533.644999999997,
+          'a22e_cper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          'a22e_pppk' => 156_082.1,
+          'a22e_pppu' => 156_082.1,
+          'a22e_ppp£' => 23_412.315000000002,
+          'a22e_ppp€' => 23_412.315000000002,
+          'a22e_pppc' => 30_107.3819,
+          'a22e_pper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          'a22e_cptr' => false,
+          'a22e_cpfa' => 9941.0,
+          'a22e_ppfa' => 9941.0,
+          'a22e_fach' => false,
+          'a22e_cpnp' => 1235,
+          'a22e_ppnp' => 1235,
+          'a22e_pnch' => false,
+          'a22e_cp£k' => 0.15,
+          'a22e_pp£k' => 0.15000000000000002,
+          'a22e_cppp' => false,
+          'a22e_ratg' => 6.8,
+          's22e_difk' => -5857.8000000000175,
+          's22e_dif£' => -878.6700000000055,
+          's22e_dif€' => -878.6700000000055,
+          's22e_difp' => -0.03753024850383239,
+          's22e_cppk' => 150_224.3,
+          's22e_cppc' => 29_528.102099999996,
+          's22e_cpp£' => 22_533.644999999997,
+          's22e_cpp€' => 22_533.644999999997,
+          's22e_cper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          's22e_pppk' => 156_082.1,
+          's22e_pppu' => 156_082.1,
+          's22e_ppp£' => 23_412.315000000002,
+          's22e_ppp€' => 23_412.315000000002,
+          's22e_pppc' => 30_107.3819,
+          's22e_pper' => 'Autumn term 2021 versus 2022 energy use comparison',
+          's22e_cptr' => false,
+          's22e_cpfa' => 9941.0,
+          's22e_ppfa' => 9941.0,
+          's22e_fach' => false,
+          's22e_cpnp' => 1235,
+          's22e_ppnp' => 1235,
+          's22e_pnch' => false,
+          's22e_cp£k' => 0.15,
+          's22e_pp£k' => 0.15000000000000002,
+          's22e_cppp' => false,
+          's22e_ratg' => 6.8,
+          'gmtc_mets' => 4,
+          'gmtc_sav£' => 2190.0,
+          'gmtc_ratg' => 0.0
+        }
+      }
+    }
+  end
 end
 
 RSpec.configure do |config|
